feat: Create the `ErrorComponent` component
## The Error Boundaries features
This first adds in utils and helpers to log the errors remotely to a third party
service or inhouse logging server. The components use the
`componentDidCatch` and `getDerivedStateFromError` class life cycle
methods to capture errors within the child UI component.
- `Note`: This might be the only places where usage of `class` components is
acceptable in the entire application. Just kidding you can do whatever you
want with your classes 😀. But seriously use functional components mostly.
- `TODO`: Add customizable fallback screens for error in simple sections like
cards or modals.

---

### feat: Create the `errorLogger` utils
This is to act as a wrapper around the external error loggers to pass error info
additional anonymous information to the logging service like
`NewRelic`/`Sentry`/`BugSnag`, `Honeybadger`, etc.
-	Add some base useful constants to uniquely identify such errors, such as
network, permissions, or UI bugs.

---

### feat: Create the `ErrorBoundaryWrapper` component
This captures error with in the underlying render method and the sub-children
of the specific component and derives a fallback screen to prevent the error
from propagating to the ancestors that might gradually crashes the app.

---

### feat: Create the `ErrorComponent` component
This component handles whenever the user encounters a JS error in the app. It
follows the "error boundary" pattern in React. We're using a class component
because according to the documentation, only class components can be error
boundaries.
- `NOTE`: Inspired by the extract from the `Ignite CLI` RN boilerplate.