feat: Create the base `Screen` components
## The Screens components
---

### feat:  Create the base `Screen` component
The central Screen component of the app, that provides a full screen view in
addition to error boundaries, out of the box.

---

### feat: Create the `useScreenSpacing` hook
This hook returns the space from the top/bottom of the screen and their
respective insets based on the device screen feature types like notches.

---

### feat: Create the `FullScreenBox` component
This creates a full screen component that takes up the entire bounds of the
screen without using the `SafeAreaView` or the template of the underlying
screen component.
-  This is to be consumed by components that want to look like screens but
don't require the whole floss of the screen.

---

### feat: Create the `Placeholder` component
This displays a for modal/screen prompting the user to update the app on
drastic feature of version change based on a remote config & store versions.